requirements.txt
python-dotenv==0.12.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
passlib==1.7.2
requests-toolbelt==0.9.1
python3-saml==1.9.0